+++ dispatcher_introspection.go
@@ -27,6 +27,7 @@ import (
 
 	tchannel "github.com/uber/tchannel-go"
 	thriftrw "go.uber.org/thriftrw/version"
+	xintrospection "go.uber.org/yarpc/api/x/introspection"
 	"go.uber.org/yarpc/internal/introspection"
 	"google.golang.org/grpc"
 )
@@ -35,23 +36,23 @@ import (
 // acquires a lots of locks throughout and should only be called with some
 // reserve.
 func (d *Dispatcher) Introspect() introspection.DispatcherStatus {
-	var inbounds []introspection.InboundStatus
+	var inbounds []xintrospection.InboundStatus
 	for _, i := range d.inbounds {
-		var status introspection.InboundStatus
-		if i, ok := i.(introspection.IntrospectableInbound); ok {
+		var status xintrospection.InboundStatus
+		if i, ok := i.(xintrospection.IntrospectableInbound); ok {
 			status = i.Introspect()
 		} else {
-			status = introspection.InboundStatus{
+			status = xintrospection.InboundStatus{
 				Transport: "Introspection not supported",
 			}
 		}
 		inbounds = append(inbounds, status)
 	}
-	var outbounds []introspection.OutboundStatus
+	var outbounds []xintrospection.OutboundStatus
 	for outboundKey, o := range d.outbounds {
 		if o.Unary != nil {
-			var status introspection.OutboundStatus
-			if o, ok := o.Unary.(introspection.IntrospectableOutbound); ok {
+			var status xintrospection.OutboundStatus
+			if o, ok := o.Unary.(xintrospection.IntrospectableOutbound); ok {
 				status = o.Introspect()
 			} else {
 				status.Transport = "Introspection not supported"
@@ -62,8 +63,8 @@ func (d *Dispatcher) Introspect() introspection.DispatcherStatus {
 			outbounds = append(outbounds, status)
 		}
 		if o.Oneway != nil {
-			var status introspection.OutboundStatus
-			if o, ok := o.Oneway.(introspection.IntrospectableOutbound); ok {
+			var status xintrospection.OutboundStatus
+			if o, ok := o.Oneway.(xintrospection.IntrospectableOutbound); ok {
 				status = o.Introspect()
 			} else {
 				status.Transport = "Introspection not supported"
@@ -97,7 +98,7 @@ var PackageVersions = []introspection.PackageVersion{
 	{Name: "go", Version: runtime.Version()},
 }
 
-type outboundStatuses []introspection.OutboundStatus
+type outboundStatuses []xintrospection.OutboundStatus
 
 func (o outboundStatuses) Len() int {
 	return len(o)